tests/submission/Infraware/Forms/contents/Forms/option_label.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_label</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1" label="label0"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var option = select.item(0);
		var label = option.label;
		option.label = 'CustomLabel'

		if (typeof(label) == "string") {
			if (label == 'label0' && option.label == 'CustomLabel' && select.item(1).label == 'Puffer') {
				test(function() {
					assert_true(true, "label attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("label attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("label attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
